requirements.txt
pandas
numpy
scikit-learn
xgboost
lightgbm
catboost
autogluon.tabular
plotly.express
seaborn